fix: moved button to its own folder
